Chatbot Bimbingan Akademik/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
mysql-connector-python==8.1.0
tensorflow==2.12.0
numpy==1.25.2
pandas==2.1.2
scikit-learn==1.2.2
PyMuPDF==1.22.3
Werkzeug==3.0.1